refactor(rc0): extract promote_topic_via_title_index + auto_select_zim_file to topic_preprocessing
