+++ openzim_mcp/simple_tools.py
@@ -29,18 +29,9 @@ from .title_promotion import (
     _DISCRIMINATOR_STOP_WORDS,
     _TAIL_TOKEN_RE,
     _TOKEN_RE,
-    accept_possessive_promotion,
-    count_non_tail_strong_entities,
     find_title_match,
     has_apostrophe_possessive,
-    has_digit_specificity_match,
-    has_topic_prefix_canonical_extension,
     is_strong_title_match,
-    is_tail_hijack_shape,
-    is_tangential_multi_token_shape,
-    iter_query_tails,
-    iter_query_windows,
-    probed_head_matches_promoted,
 )
 from .tool_schemas import (
     SearchAllResponse,
@@ -3896,164 +3887,21 @@ class SimpleToolsHandler:
     def _promote_topic_via_title_index(
         self, zim_file_path: str, topic: str
     ) -> Optional[Dict[str, Any]]:
-        """Resolve ``topic`` to a canonical title-index hit.
+        """Phase F: thin wrapper delegating to ``topic_preprocessing``.
 
-        Pass order (first hit wins):
-
-          1. Strict 1.0 gate over the trailing tails
-             (``iter_query_tails``) — preserves a14's motivating
-             "famous people from big rapids michigan" behavior, where
-             the entity sits at the tail of the prose.
-          2. Strict 1.0 gate over non-trailing sliding windows
-             (``iter_query_windows``) — post-a14 sweep (F4 / A2):
-             catches queries whose entity sits at the head/middle
-             ("Big Rapids Michigan tourism"). Only fires when no
-             trailing tail resolved strictly, so existing behavior is
-             unchanged for tail-positioned entities.
-          3. Typo-tolerant 0.8 gate over the trailing tails — catches
-             single-edit typos like ``Photosythesis`` →
-             ``Photosynthesis``.
-
-        D6 fix (v2.0.0a9): Xapian ranks ``List of songs about Berlin``
-        above the canonical ``Berlin`` article for ``query=Berlin``
-        because the title-match boost isn't strong enough. The 1.0 gate
-        promotes the canonical past that ranking.
+        See :func:`openzim_mcp.topic_preprocessing.promote_topic_via_title_index`
+        for the Z3/Z4/OPP-1 pass-order documentation. Behavior is byte-identical
+        to the pre-extraction implementation (proven by the Gate 0a parity
+        diff-test in ``tests/dispatch_eval/test_promotion_extraction_parity.py``
+        replaying the b1→b13 probe set against a live Wikipedia ZIM).
         """
-        # Pass 0: probe the FULL topic (with original punctuation
-        # preserved) BEFORE iter_query_tails decomposes it.
-        # ``iter_query_tails`` tokenizes on alphanumeric runs, so
-        # apostrophe-bearing possessives split: ``einstein's theory``
-        # becomes the tokens ``["einstein", "s", "theory"]`` and the
-        # yielded tails (``"einstein s theory"`` / ``"s theory"`` /
-        # ``"theory"``) all lose the apostrophe. The longer tails
-        # then fail to match any canonical title-index entry (which
-        # are stored WITH the apostrophe — ``Einstein's_theory``
-        # redirects to ``Theory_of_relativity``), and the shortest
-        # tail ``"theory"`` wins via the generic ``Theory`` article.
-        # Probing the original topic with punctuation at the start
-        # catches these. ``min_score=0.95`` mirrors the canonical-or-
-        # fuzzy gate Rule 2/3/4's probe uses (intent_parser.py:317),
-        # accepting both direct hits (1.0) and high-confidence
-        # redirects / spelling variants (0.95). The b3 invariant
-        # (regression-guarded by test_post_b3 / test_promote_function_
-        # starts_with_full_topic_probe) requires this to be the FIRST
-        # ``find_title_match`` call in the function.
-        promoted = find_title_match(
-            self.zim_operations, zim_file_path, topic, min_score=0.95
+        from openzim_mcp.topic_preprocessing import promote_topic_via_title_index
+
+        return promote_topic_via_title_index(
+            zim_operations=self.zim_operations,
+            zim_file_path=zim_file_path,
+            topic=topic,
         )
-        # Post-b4 D2: when the topic carries an apostrophe-possessive
-        # (``Plato's republic philosophy`` / ``Einstein's theory
-        # history``), tighten the tail-iteration floor to ``min_len=2``
-        # so a generic 1-token trailing tail (``"philosophy"`` /
-        # ``"history"`` / ``"tourism"``) can't silently win at strict
-        # 1.0. Pass-0 already covered the legitimate "full X's Y is
-        # canonical" case; further 1-token tail iteration just risks
-        # picking a generic Wikipedia title that shares the word.
-        pass_tail_min_len = 2 if has_apostrophe_possessive(topic) else 1
-
-        # Token probe shared by the b10 Z3 multi-entity discriminator
-        # AND the b11 Z4 head-biographical-canonical check. Wrapped
-        # once and reused so the cost stays bounded at one libzim probe
-        # per non-tail topic token (and one per head check).
-        def _probe(token_str: str) -> Optional[Dict[str, Any]]:
-            return find_title_match(self.zim_operations, zim_file_path, token_str)
-
-        # Post-b11 Z4 layer. Multi-token tangential canonicals
-        # (``Lenin Russia`` → ``Leninist_Komsomol_of_the_Russian_Federation``,
-        # ``Tesla electricity`` → ``Tesla's_Wireless_Electricity``,
-        # ``Mozart Vienna`` → ``Mozarthaus_Vienna``,
-        # ``Beethoven symphony`` → ``Symphony_No._1_(Beethoven)``,
-        # ``Marie Curie radioactivity`` → ``Radioactive_(Redniss_book)``)
-        # surface where ``is_tail_hijack_shape`` doesn't fire (canonical
-        # is multi-token). Z4 rejects them UNLESS one of three
-        # exemptions applies: biographical-canonical (head probe
-        # matches promoted), digit-specificity (numbered-instance
-        # signal), or type-extension (canonical's leading tokens are
-        # a 2+-token contiguous topic slice with extras only at
-        # suffix — preserves ``Big Rapids Michigan Ferris State`` →
-        # ``Ferris_State_University``). The Z4 check is applied at
-        # every pass (0, 1, 2, 3) so the defect class can't slip
-        # through a different gate.
-        def _passes_z4(promoted_arg: Dict[str, Any]) -> bool:
-            if has_apostrophe_possessive(topic):
-                return True
-            if not is_tangential_multi_token_shape(promoted_arg, topic):
-                return True
-            if probed_head_matches_promoted(topic, promoted_arg, _probe):
-                return True
-            if has_digit_specificity_match(promoted_arg, topic):
-                return True
-            if has_topic_prefix_canonical_extension(promoted_arg, topic):
-                return True
-            return False
-
-        # Post-b10 Z3 multi-entity discriminator wrapper. Pass 1 / Pass 2
-        # consult this gate, which layers the b10 single-entity escape
-        # over the b9 unconditional tail-hijack rejection, then applies
-        # the b11 Z4 check on top. Pass 0 / Pass 3 use the bare
-        # ``accept_possessive_promotion`` + ``_passes_z4`` pair (no
-        # single-entity escape — that escape exists for Pass 1's
-        # 1-token-tail filler-prose pattern, not for Pass 0's full-
-        # topic probe).
-        def _accept_with_multi_entity_check(
-            promoted_arg: Dict[str, Any],
-        ) -> bool:
-            if has_apostrophe_possessive(topic):
-                return accept_possessive_promotion(promoted_arg, topic)
-            base_accept = accept_possessive_promotion(promoted_arg, topic)
-            if not base_accept:
-                if not is_tail_hijack_shape(promoted_arg, topic):
-                    return False
-                return count_non_tail_strong_entities(topic, _probe, limit=2) < 2
-            return _passes_z4(promoted_arg)
-
-        # Pass 0 acceptance: base accept gate + Z4 layer. Stalin USSR
-        # Russia → Russia (tail-hijack) is rejected here unconditionally
-        # because ``accept_possessive_promotion`` already says False —
-        # the multi-entity escape only fires at Pass 1/2. Tesla
-        # electricity → Tesla's_Wireless_Electricity is accepted by
-        # ``accept_possessive_promotion`` (2-token topic, no tail-
-        # hijack) but rejected by ``_passes_z4`` (tangential shape,
-        # head probe differs from promoted).
-        if (
-            promoted is not None
-            and accept_possessive_promotion(promoted, topic)
-            and _passes_z4(promoted)
-        ):
-            return promoted
-
-        # Pass 1: strict 1.0-score gate across every trailing tail.
-        # Prefer an exact title match on any tail (even a short one)
-        # over a typo-tolerant fuzzy match on a longer noisier tail.
-        for tail in iter_query_tails(topic, min_len=pass_tail_min_len):
-            promoted = find_title_match(self.zim_operations, zim_file_path, tail)
-            if promoted is not None and _accept_with_multi_entity_check(promoted):
-                return promoted
-        # Pass 2: strict 1.0-score gate across non-trailing windows.
-        # Catches head/middle-positioned entities like
-        # ``Big Rapids Michigan tourism``. Length-decreasing so longer
-        # (more specific) windows win.
-        for window in iter_query_windows(topic, min_len=pass_tail_min_len):
-            promoted = find_title_match(self.zim_operations, zim_file_path, window)
-            if promoted is not None and _accept_with_multi_entity_check(promoted):
-                return promoted
-        # Pass 3: 0.8 typo-tolerant gate. Only fires when no strict
-        # match exists at all — catches single-edit typos. The bare
-        # ``accept_possessive_promotion`` + ``_passes_z4`` pair (no
-        # multi-entity escape) is the same shape as Pass 0; the
-        # escape's filler-prose pattern doesn't apply to typo-corrected
-        # tails.
-        for tail in iter_query_tails(topic, min_len=pass_tail_min_len):
-            promoted = find_title_match(
-                self.zim_operations, zim_file_path, tail, min_score=0.8
-            )
-            if (
-                promoted is not None
-                and accept_possessive_promotion(promoted, topic)
-                and _passes_z4(promoted)
-            ):
-                return promoted
-        return None
 
     def _handle_tell_me_about(
         self,
@@ -5782,36 +5630,14 @@ class SimpleToolsHandler:
             return None
 
     def _auto_select_zim_file(self) -> Optional[str]:
-        """Auto-select a ZIM file if only one is available.
+        """Phase F: thin wrapper delegating to ``topic_preprocessing``.
 
-        Returns:
-            Path to ZIM file if exactly one exists, None otherwise.
-            Returns None with appropriate logging if multiple files exist
-            or on error.
+        See :func:`openzim_mcp.topic_preprocessing.auto_select_zim_file` for
+        the 4-arm log envelope (0-files info / 1-file debug / N-files info /
+        exception warning). Byte-identical to the pre-extraction implementation
+        (proven by the Gate 0a parity diff-test in
+        ``tests/dispatch_eval/test_auto_select_extraction_parity.py``).
         """
-        try:
-            # Use structured data method directly (not parsing JSON from string)
-            files = self.zim_operations.list_zim_files_data()
+        from openzim_mcp.topic_preprocessing import auto_select_zim_file
 
-            if len(files) == 0:
-                logger.info(
-                    "Auto-select failed: no ZIM files found in allowed directories"
-                )
-                return None
-            elif len(files) == 1:
-                selected = str(files[0]["path"])
-                logger.debug(f"Auto-selected ZIM file: {selected}")
-                return selected
-            else:
-                logger.info(
-                    f"Auto-select skipped: {len(files)} ZIM files found, "
-                    "please specify which file to use"
-                )
-                return None
-
-        except Exception as e:
-            # Log at warning level with specific error for debugging
-            logger.warning(
-                f"Auto-select ZIM file failed with error: {type(e).__name__}: {e}"
-            )
-            return None
+        return auto_select_zim_file(self.zim_operations)
