+++ tests/test_post_b4_beta_fixes.py
@@ -218,7 +218,7 @@ class TestFuzzySuggestGateReject:
                 }
             return None
 
-        with patch("openzim_mcp.simple_tools.find_title_match", side_effect=fake):
+        with patch("openzim_mcp.topic_preprocessing.find_title_match", side_effect=fake):
             result = SimpleToolsHandler._promote_topic_via_title_index(
                 self._make_handler(),
                 "test.zim",
@@ -264,7 +264,7 @@ class TestFuzzySuggestGateReject:
                 }
             return None
 
-        with patch("openzim_mcp.simple_tools.find_title_match", side_effect=fake):
+        with patch("openzim_mcp.topic_preprocessing.find_title_match", side_effect=fake):
             result = SimpleToolsHandler._promote_topic_via_title_index(
                 self._make_handler(),
                 "test.zim",
@@ -319,7 +319,7 @@ class TestFuzzySuggestGateReject:
                 }
             return None
 
-        with patch("openzim_mcp.simple_tools.find_title_match", side_effect=fake):
+        with patch("openzim_mcp.topic_preprocessing.find_title_match", side_effect=fake):
             result = SimpleToolsHandler._promote_topic_via_title_index(
                 self._make_handler(),
                 "test.zim",
@@ -355,7 +355,7 @@ class TestFuzzySuggestGateReject:
                 }
             return None
 
-        with patch("openzim_mcp.simple_tools.find_title_match", side_effect=fake):
+        with patch("openzim_mcp.topic_preprocessing.find_title_match", side_effect=fake):
             result = SimpleToolsHandler._promote_topic_via_title_index(
                 self._make_handler(),
                 "test.zim",
@@ -468,7 +468,7 @@ class TestPossessiveMinLenFloor:
                 }
             return None
 
-        with patch("openzim_mcp.simple_tools.find_title_match", side_effect=fake):
+        with patch("openzim_mcp.topic_preprocessing.find_title_match", side_effect=fake):
             result = SimpleToolsHandler._promote_topic_via_title_index(
                 self._make_handler(),
                 "test.zim",
@@ -514,7 +514,7 @@ class TestPossessiveMinLenFloor:
                 }
             return None
 
-        with patch("openzim_mcp.simple_tools.find_title_match", side_effect=fake):
+        with patch("openzim_mcp.topic_preprocessing.find_title_match", side_effect=fake):
             result = SimpleToolsHandler._promote_topic_via_title_index(
                 self._make_handler(),
                 "test.zim",
@@ -551,7 +551,7 @@ class TestPossessiveMinLenFloor:
                 }
             return None
 
-        with patch("openzim_mcp.simple_tools.find_title_match", side_effect=fake):
+        with patch("openzim_mcp.topic_preprocessing.find_title_match", side_effect=fake):
             result = SimpleToolsHandler._promote_topic_via_title_index(
                 self._make_handler(),
                 "test.zim",
